pyproject.toml
[project]
name = "gerrit-cli"
version = "0.1.0"
description = "A command-line interface for Gerrit Code Review"
readme = "README.md"
authors = [
    { name = "madroid", email = "madroidmaq@gmail.com" }
]
requires-python = ">=3.10"
dependencies = [
    "click>=8.1.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "rich>=13.0.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
gerrit = "gerrit_cli.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "respx>=0.21.0",
    "black>=24.0.0",
    "ruff>=0.6.0",
    "mypy>=1.11.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 100

[tool.ruff]
line-length = 100

[tool.mypy]
python_version = "3.10"
strict = true
